% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpointing.R
\name{checkpoint}
\alias{checkpoint}
\title{Checkpoint at current or given sequence number}
\usage{
checkpoint(sequenceNumber)
}
\arguments{
\item{sequenceNumber}{optional}
}
\description{
Checkpoint at current or given sequence number
}
