% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{kineRic-package}
\alias{kineRic-package}
\title{An R Kinesis Consumer}
\description{
This is a placeholder for storing import directives, please find more details in the \code{README.md} file.
}

