% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kineric.R
\name{kineric}
\alias{kineric}
\title{Run Kinesis Consumer application}
\usage{
kineric(initialize, processRecords, shutdown, checkpointing = TRUE,
  logfile = tempfile())
}
\arguments{
\item{initialize}{optional function to be run on startup}

\item{processRecords}{function to process records taking a \code{data.frame} object with \code{partitionKey}, \code{sequenceNumber} and \code{data} columns as the \code{records} argument. Probably you only need the \code{data} column from this object}

\item{shutdown}{optional function to be run when finished processing all records in a shard}

\item{checkpointing}{if set to \code{TRUE} (default), \code{kineric} will checkpoint after each \code{processRecords} call. To disable checkpointing altogether, set this to \code{FALSE}. If you want to checkpoint periodically, set this to the frequency in minutes as integer.}

\item{logfile}{file path of the log file. To disable logging, set \code{flog.threshold} to something high}
}
\description{
Run Kinesis Consumer application
}

