% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kms.R
\name{kms_decrypt}
\alias{kms_decrypt}
\title{Decrypt cipher into plain text via KMS}
\usage{
kms_decrypt(cipher)
}
\arguments{
\item{cipher}{base64-encoded ciphertext}
}
\value{
string
}
\description{
Decrypt cipher into plain text via KMS
}

