% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kms.R
\name{kms_encrypt}
\alias{kms_encrypt}
\title{Encrypt plain text via KMS}
\usage{
kms_encrypt(key, text)
}
\arguments{
\item{key}{KMS key ID, fully specified Amazon Resource Name or an alias with the \code{alias/} prefix}

\item{text}{max 4096 bytes long string}
}
\value{
base64-encoded text
}
\description{
Encrypt plain text via KMS
}
\examples{
\dontrun{
kms_encrypt('alias/mykey', 'foobar')
}
}

