% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_line_from_stdin}
\alias{read_line_from_stdin}
\title{Read one non-empty line from stdin without any warnings printed to stdout}
\usage{
read_line_from_stdin()
}
\value{
string
}
\description{
Read one non-empty line from stdin without any warnings printed to stdout
}
\keyword{internal}
