% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_line_to_stdout}
\alias{write_line_to_stdout}
\title{Securely write a line to stdout with logging}
\usage{
write_line_to_stdout(line)
}
\arguments{
\item{line}{string}
}
\description{
Securely write a line to stdout with logging
}
\keyword{internal}
