% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{kinesis_get_records}
\alias{kinesis_get_records}
\title{Get record from a Kinesis Stream}
\usage{
kinesis_get_records(stream, region = "us-west-1", limit = 25)
}
\arguments{
\item{stream}{stream name (string)}

\item{region}{AWS region (string)}

\item{limit}{number of records to fetch}
}
\value{
character vector that you might want to post-process eg with \code{jsonlite::stream_in}
}
\description{
Get record from a Kinesis Stream
}
\note{
Use this no more than getting sample data from a stream - it's not intended for prod usage.
}
\references{
\url{https://docs.aws.amazon.com/AWSJavaSDK/latest/javadoc/com/amazonaws/services/kinesis/model/GetRecordsRequest.html}
}
